feat: add spawner injection seam to cron-handler preCheck (CPS-1.1)
preCheck's Bun.spawn(["bun", scriptPath]) call was hardcoded with no
injection seam, forcing cron-handler.smoke.test.ts to spawn ~15 real bun
child processes per test run — a confirmed source of intermittent CI
failures under resource contention (PR #3006 / SSB-1.1).

Add an optional `spawner` field to CronHandlerDeps (defaults to real
Bun.spawn) mirroring claude.ts's existing spawner pattern, and use
`deps.spawner ?? Bun.spawn` at the preCheck call site. Production
behavior is unchanged.

Relocate the ~15 preCheck-behavior tests out of cron-handler.smoke.test.ts
(smoke tests are HTTP route contracts per this repo's test-layer
convention, not subprocess-spawning behavior) into a new
cron-handler.unit.test.ts, rewritten to inject a fake spawner returning
canned stdout/stderr/exit code — mirroring claude.unit.test.ts's
fakeProc/mockSpawn pattern — instead of spawning real bun processes.
Keep exactly one true end-to-end test in cron-handler.integration.test.ts
that omits the spawner override to exercise the real default-fallback
wiring.
