+++ agent/src/cron-handler.smoke.test.ts
@@ -22,10 +22,10 @@ import { join } from "node:path";
 import type { ErrorCapturingClient } from "@shipwright/lib/sentry";
 import type { WebClient } from "@slack/web-api";
 import type { Server } from "bun";
-import { FixedClock } from "./clock.ts";
 import type { ModelUsage, TokenUsage } from "./claude.ts";
-import { ValidationError, handleCronRequest } from "./cron-handler.ts";
+import { FixedClock } from "./clock.ts";
 import { reportCronFailure } from "./cron-failure-reporter.ts";
+import { ValidationError, handleCronRequest } from "./cron-handler.ts";
 import { HttpCronRunReporter } from "./cron-run-reporter.ts";
 import { startHealthServer } from "./health.ts";
 
@@ -551,429 +551,6 @@ describe("handleCronRequest — runner error", () => {
   });
 });
 
-// ─── preCheck ─────────────────────────────────────────────────────────────────
-
-describe("handleCronRequest — preCheck", () => {
-  let tmpDir: string;
-
-  beforeEach(() => {
-    tmpDir = mkdtempSync(join(tmpdir(), "cron-test-"));
-    mkdirSync(join(tmpDir, "shipwright", "scripts"), { recursive: true });
-  });
-
-  test("skips job when preCheck script exits 1 (no work)", async () => {
-    const script = join(tmpDir, "check.ts");
-    writeFileSync(script, "process.exit(1);");
-
-    await handleCronRequest(
-      { jobId: "j1", prompt: "hello", channel: "C-TEST", preCheck: script },
-      deps,
-    );
-
-    expect(mockRunner).not.toHaveBeenCalled();
-    expect(mockPostMessage).not.toHaveBeenCalled();
-  });
-
-  test("uses preCheck output as prompt when it exits 0 with output", async () => {
-    const script = join(tmpDir, "check.ts");
-    writeFileSync(script, `console.log("precheck prompt"); process.exit(0);`);
-
-    await handleCronRequest(
-      { jobId: "j1", prompt: "original", channel: "C-TEST", preCheck: script },
-      deps,
-    );
-
-    expect(mockRunner).toHaveBeenCalledTimes(1);
-    const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-    expect(runArg).toContain("precheck prompt");
-  });
-
-  test("preCheck inherits runtime process.env mutations (env: process.env passthrough)", async () => {
-    // config-sync mutates process.env at runtime (Object.assign every 60s). The
-    // preCheck child must see those mutations, not a Bun-startup snapshot —
-    // regression for a rotated SHIPWRIGHT_TASK_STORE_TOKEN causing 401s until
-    // the pod restarted. The probe key is absent from the process's boot env, so
-    // without env: process.env the child reads "MISSING".
-    const key = "CRON_PRECHECK_ENV_PROBE";
-    process.env[key] = "runtime-value";
-    try {
-      const script = join(tmpDir, "check.ts");
-      writeFileSync(
-        script,
-        `console.log(process.env.${key} ?? "MISSING"); process.exit(0);`,
-      );
-
-      await handleCronRequest(
-        {
-          jobId: "j1",
-          prompt: "original",
-          channel: "C-TEST",
-          preCheck: script,
-        },
-        deps,
-      );
-
-      expect(mockRunner).toHaveBeenCalledTimes(1);
-      const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-      expect(runArg).toContain("runtime-value");
-    } finally {
-      delete process.env[key];
-    }
-  });
-
-  test("runs preCheck with cwd = workspace (resolves state relative to workspace)", async () => {
-    // Plugin preChecks resolve state relative to process.cwd() — e.g.
-    // check-review.ts / check-deploy.ts (via check-helpers.ts) read
-    // workspace-relative files like `state/agent-policy.md`. The file lives
-    // in the workspace; the preCheck must run there, not in the agent's cwd
-    // (/app in prod, the repo root in tests).
-    const ws = join(tmpDir, "ws");
-    mkdirSync(join(ws, "state"), { recursive: true });
-    writeFileSync(join(ws, "state", "agent-policy.md"), "# Agent Policy\n");
-    const script = join(tmpDir, "check.ts");
-    writeFileSync(
-      script,
-      `import { existsSync } from "node:fs";\nconsole.log(existsSync("state/agent-policy.md") ? "HAS_STATE" : "NO_STATE");\nprocess.exit(0);`,
-    );
-
-    await handleCronRequest(
-      { jobId: "j1", prompt: "original", channel: "C-TEST", preCheck: script },
-      { ...deps, workspace: ws },
-    );
-
-    expect(mockRunner).toHaveBeenCalledTimes(1);
-    const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-    expect(runArg).toContain("HAS_STATE");
-  });
-
-  test("skips job when preCheck script not found (plugin: format)", async () => {
-    await handleCronRequest(
-      {
-        jobId: "j1",
-        prompt: "hello",
-        channel: "C-TEST",
-        preCheck: "nonexistent-plugin:check.ts",
-      },
-      { ...deps, pluginCacheDir: tmpDir },
-    );
-
-    expect(mockRunner).not.toHaveBeenCalled();
-  });
-
-  test("preCheck exits 2 → session skipped, runner NOT called, alert posted to alertsChannel", async () => {
-    const scriptPath = join(
-      tmpDir,
-      "shipwright",
-      "scripts",
-      "check-dev-task.ts",
-    );
-    writeFileSync(
-      scriptPath,
-      `#!/usr/bin/env bun\nprocess.stderr.write("something blew up\\n");\nprocess.exit(2);\n`,
-      { mode: 0o755 },
-    );
-
-    await handleCronRequest(
-      {
-        jobId: "precheck-crash",
-        prompt: "original prompt",
-        silent: true,
-        preCheck: "shipwright:check-dev-task.ts",
-      },
-      { ...deps, pluginCacheDir: tmpDir, alertsChannel: "C_ALERTS" },
-    );
-
-    expect(mockRunner).not.toHaveBeenCalled();
-    expect(mockPostMessage).toHaveBeenCalledTimes(1);
-    const call = (
-      mockPostMessage.mock.calls as unknown as unknown[][]
-    )[0][0] as {
-      channel: string;
-      text: string;
-    };
-    expect(call.channel).toBe("C_ALERTS");
-    expect(call.text).toContain("precheck-crash");
-    expect(call.text).toContain("exit 2");
-  });
-
-  test("preCheck exits 2 with no alertsChannel → session skipped, no alert, no throw", async () => {
-    const scriptPath = join(
-      tmpDir,
-      "shipwright",
-      "scripts",
-      "check-dev-task.ts",
-    );
-    writeFileSync(scriptPath, "#!/usr/bin/env bun\nprocess.exit(2);\n", {
-      mode: 0o755,
-    });
-
-    await expect(
-      handleCronRequest(
-        {
-          jobId: "precheck-crash-no-alert",
-          prompt: "original prompt",
-          silent: true,
-          preCheck: "shipwright:check-dev-task.ts",
-        },
-        { ...deps, pluginCacheDir: tmpDir },
-      ),
-    ).resolves.toBeUndefined();
-
-    expect(mockRunner).not.toHaveBeenCalled();
-    expect(mockPostMessage).not.toHaveBeenCalled();
-  });
-
-  test("preCheck exits 0 with no output → session skipped, runner NOT called", async () => {
-    const scriptPath = join(
-      tmpDir,
-      "shipwright",
-      "scripts",
-      "check-dev-task.ts",
-    );
-    writeFileSync(scriptPath, "#!/usr/bin/env bun\nprocess.exit(0);\n", {
-      mode: 0o755,
-    });
-
-    await handleCronRequest(
-      {
-        jobId: "precheck-no-output",
-        prompt: "original prompt",
-        silent: true,
-        preCheck: "shipwright:check-dev-task.ts",
-      },
-      { ...deps, pluginCacheDir: tmpDir },
-    );
-
-    expect(mockRunner).not.toHaveBeenCalled();
-  });
-
-  test("preCheck script not found → warning logged, runner NOT called, no throw", async () => {
-    const warnMessages: string[] = [];
-    const originalWarn = console.warn;
-    console.warn = (...args: unknown[]) => {
-      warnMessages.push(args.map(String).join(" "));
-    };
-
-    try {
-      await expect(
-        handleCronRequest(
-          {
-            jobId: "precheck-missing",
-            prompt: "original prompt",
-            silent: true,
-            preCheck: "shipwright:check-dev-task.ts",
-          },
-          { ...deps, pluginCacheDir: tmpDir },
-        ),
-      ).resolves.toBeUndefined();
-
-      expect(mockRunner).not.toHaveBeenCalled();
-      expect(
-        warnMessages.some((m) => m.includes("preCheck script not found")),
-      ).toBe(true);
-    } finally {
-      console.warn = originalWarn;
-    }
-  });
-
-  test("preCheck resolves via installed_plugins.json manifest (production path)", async () => {
-    const installDir = mkdtempSync(join(tmpdir(), "cron-precheck-install-"));
-    const installPath = join(installDir, "shipwright");
-    mkdirSync(join(installPath, "scripts"), { recursive: true });
-
-    const scriptPath = join(installPath, "scripts", "check-dev-task.ts");
-    writeFileSync(
-      scriptPath,
-      `#!/usr/bin/env bun\nconsole.log("manifest path works");\nprocess.exit(0);\n`,
-      { mode: 0o755 },
-    );
-
-    const manifestDir = mkdtempSync(join(tmpdir(), "cron-precheck-manifest-"));
-    const manifestPath = join(manifestDir, "installed_plugins.json");
-    writeFileSync(
-      manifestPath,
-      JSON.stringify({
-        version: 2,
-        plugins: {
-          "shipwright@app-vitals/shipwright": [{ installPath }],
-        },
-      }),
-    );
-
-    // biome-ignore lint/suspicious/noExplicitAny: test mock — sessionId optional
-    mockRunner.mockResolvedValueOnce({ result: "done [silent]" } as any);
-
-    await handleCronRequest(
-      {
-        jobId: "precheck-manifest",
-        prompt: "original prompt",
-        silent: true,
-        preCheck: "shipwright:check-dev-task.ts",
-      },
-      { ...deps, pluginManifestPath: manifestPath },
-    );
-
-    expect(mockRunner).toHaveBeenCalledTimes(1);
-    const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-    expect(runArg).toContain("manifest path works");
-  });
-
-  test("preCheck manifest path: missing plugins.json → warning logged, runner NOT called", async () => {
-    const warnMessages: string[] = [];
-    const originalWarn = console.warn;
-    console.warn = (...args: unknown[]) => {
-      warnMessages.push(args.map(String).join(" "));
-    };
-
-    try {
-      await expect(
-        handleCronRequest(
-          {
-            jobId: "precheck-nomanifest",
-            prompt: "original prompt",
-            silent: true,
-            preCheck: "shipwright:check-dev-task.ts",
-          },
-          {
-            ...deps,
-            pluginManifestPath: join(
-              tmpDir,
-              "nonexistent_installed_plugins.json",
-            ),
-          },
-        ),
-      ).resolves.toBeUndefined();
-
-      expect(mockRunner).not.toHaveBeenCalled();
-      expect(
-        warnMessages.some(
-          (m) =>
-            m.includes("failed to read installed_plugins.json") ||
-            m.includes("preCheck script not found"),
-        ),
-      ).toBe(true);
-    } finally {
-      console.warn = originalWarn;
-    }
-  });
-
-  test("relative path (./scripts/check.ts) resolves against workspace, output becomes prompt", async () => {
-    const tmpWorkspace = mkdtempSync(join(tmpdir(), "cron-precheck-relpath-"));
-    mkdirSync(join(tmpWorkspace, "scripts"), { recursive: true });
-    const scriptPath = join(tmpWorkspace, "scripts", "check.ts");
-    writeFileSync(
-      scriptPath,
-      `#!/usr/bin/env bun\nconsole.log("workspace task found");\nprocess.exit(0);\n`,
-      { mode: 0o755 },
-    );
-
-    // biome-ignore lint/suspicious/noExplicitAny: test mock — sessionId optional
-    mockRunner.mockResolvedValueOnce({ result: "done [silent]" } as any);
-
-    await handleCronRequest(
-      {
-        jobId: "precheck-relpath",
-        prompt: "original prompt",
-        silent: true,
-        preCheck: "./scripts/check.ts",
-      },
-      { ...deps, workspace: tmpWorkspace },
-    );
-
-    expect(mockRunner).toHaveBeenCalledTimes(1);
-    const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-    expect(runArg).toContain("workspace task found");
-  });
-
-  test("absolute path resolves directly regardless of workspace", async () => {
-    const absDir = mkdtempSync(join(tmpdir(), "cron-precheck-abs-"));
-    const scriptPath = join(absDir, "absolute-check.ts");
-    writeFileSync(
-      scriptPath,
-      `#!/usr/bin/env bun\nconsole.log("absolute path works");\nprocess.exit(0);\n`,
-      { mode: 0o755 },
-    );
-
-    // biome-ignore lint/suspicious/noExplicitAny: test mock — sessionId optional
-    mockRunner.mockResolvedValueOnce({ result: "done [silent]" } as any);
-
-    await handleCronRequest(
-      {
-        jobId: "precheck-absolute",
-        prompt: "original prompt",
-        silent: true,
-        preCheck: scriptPath,
-      },
-      { ...deps },
-    );
-
-    expect(mockRunner).toHaveBeenCalledTimes(1);
-    const runArg = (mockRunner.mock.calls as unknown as string[][])[0][0];
-    expect(runArg).toContain("absolute path works");
-  });
-
-  test("relative path without workspace → warning, runner NOT called", async () => {
-    const warnMessages: string[] = [];
-    const originalWarn = console.warn;
-    console.warn = (...args: unknown[]) => {
-      warnMessages.push(args.map(String).join(" "));
-    };
-
-    try {
-      await expect(
-        handleCronRequest(
-          {
-            jobId: "precheck-relpath-noworkspace",
-            prompt: "original prompt",
-            silent: true,
-            preCheck: "./scripts/check.ts",
-          },
-          { ...deps },
-        ),
-      ).resolves.toBeUndefined();
-
-      expect(mockRunner).not.toHaveBeenCalled();
-      expect(
-        warnMessages.some(
-          (m) => m.includes("relative") || m.includes("workspace"),
-        ),
-      ).toBe(true);
-    } finally {
-      console.warn = originalWarn;
-    }
-  });
-
-  test("relative path script not found → warning, runner NOT called", async () => {
-    const tmpWorkspace = mkdtempSync(join(tmpdir(), "cron-precheck-notfound-"));
-    const warnMessages: string[] = [];
-    const originalWarn = console.warn;
-    console.warn = (...args: unknown[]) => {
-      warnMessages.push(args.map(String).join(" "));
-    };
-
-    try {
-      await expect(
-        handleCronRequest(
-          {
-            jobId: "precheck-relpath-notfound",
-            prompt: "original prompt",
-            silent: true,
-            preCheck: "./nonexistent.ts",
-          },
-          { ...deps, workspace: tmpWorkspace },
-        ),
-      ).resolves.toBeUndefined();
-
-      expect(mockRunner).not.toHaveBeenCalled();
-      expect(
-        warnMessages.some((m) => m.includes("preCheck script not found")),
-      ).toBe(true);
-    } finally {
-      console.warn = originalWarn;
-    }
-  });
-});
-
 // ─── CronRunReporter integration ─────────────────────────────────────────────
 
 const FIXED_TIME = new Date("2026-01-01T00:00:00.000Z");
@@ -1072,7 +649,10 @@ describe("handleCronRequest — CronRunReporter", () => {
     expect(reporterState.requests[0].url).toContain(
       `/agents/${AGENT_ID}/crons/create-test/runs`,
     );
-    const createBody = reporterState.requests[0].body as Record<string, unknown>;
+    const createBody = reporterState.requests[0].body as Record<
+      string,
+      unknown
+    >;
     expect(createBody.startedAt).toBe(FIXED_TIME.toISOString());
     expect(Object.keys(createBody)).toEqual(["startedAt"]);
   });
@@ -1170,9 +750,7 @@ describe("handleCronRequest — CronRunReporter", () => {
       Record<string, unknown>
     >;
     expect(modelBreakdown).toHaveLength(2);
-    const byModel = Object.fromEntries(
-      modelBreakdown.map((m) => [m.model, m]),
-    );
+    const byModel = Object.fromEntries(modelBreakdown.map((m) => [m.model, m]));
     expect(byModel["claude-sonnet-4-6"].costUsd).toBe(0.002);
     expect(byModel["claude-opus-4-8"].costUsd).toBe(0.01);
   });
